Update on the solution of 704. Binary Search to avoid overflow

+++ Array/704. Binary Search/main.cpp
@@ -12,7 +12,7 @@ public:
 		if (n == 0 || n == 1 && nums.at(0) != target) return -1;
 		int left = 0, right = n - 1;
 		while (left <= right) {
-			int mid = (left + right) >> 1; // To avoid overflow
+			int mid = (unsigned int)(left + right) >> 1; // To avoid overflow
 			if (nums.at(mid) == target) return mid;
 			else if (nums.at(mid) < target) left = mid + 1;
 			else right = mid - 1;
